requirements.txt
supabase-py==2.4.0
